Moved redundant Id64 trait impls into macro
Again, the new generate_core_trait_impls macro can be used with future Id32 and Id16 types

The `impl fmt::Display for Id64` has deliberately not been included, as implementing a variable number of input params within the write! call seems impratical at this time. Further work will need to refactor this impl to make it easier to port over to the macro.